+++ changelog.md
@@ -1,6 +1,7 @@
 # Changelog
 
-- * 2620093 - 2018-10-27: Release v9.4.1 
+- * f1b4113 - 2018-10-28: Added function to strip accents to encoding functions 
+- * bf93454 - 2018-10-27: Release v9.4.1 
 - * 3dbd257 - 2018-10-27: Release v9.3.100 
 - * 479d80e - 2018-10-27: Removed app_id and replaced with app_sub. App now links its self to SSO user via userid instead of appid. SSL cert and key now stored/retrieved from data folder. 
 - * 33ea2f1 - 2018-10-21: Release v9.3.99